Refactor Abstract*VersionStore* test classes (PR 1/2)
The base test class for the VersionStore became pretty big, so it's worth to
move groups of tests to nested test classes.

This PR prepares the spill-out to nested test classes.

* Rename `AbstractITVersionStore` to `AbstractVersionStoreTestBase`
* Rename `AbstractVersionStoreTest` to `AbstractDatabaseAdapterVersionStoreTest`
* Move non-database-adapter specific tests to `AbstractVersionStoreTestBase`
