+++ versioned/tests/src/main/java/org/projectnessie/versioned/tests/AbstractVersionStoreTestBase.java
@@ -15,6 +15,7 @@
  */
 package org.projectnessie.versioned.tests;
 
+import static java.util.Collections.singletonList;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -39,6 +40,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
 import org.projectnessie.versioned.BranchName;
 import org.projectnessie.versioned.Commit;
 import org.projectnessie.versioned.Delete;
@@ -61,7 +64,7 @@ import org.projectnessie.versioned.WithHash;
 import org.projectnessie.versioned.WithType;
 
 /** Base class used for integration tests against version store implementations. */
-public abstract class AbstractITVersionStore {
+public abstract class AbstractVersionStoreTestBase {
 
   protected abstract VersionStore<String, String, StringStoreWorker.TestEnum> store();
 
@@ -1477,5 +1480,281 @@ public abstract class AbstractITVersionStore {
 
   protected static Commit<String, String> commit(Hash hash, String commitMeta) {
     return Commit.<String, String>builder().hash(hash).commitMeta(commitMeta).build();
+  }
+
+  /** Test behavior when a table (content-key) is created, then dropped and then re-created. */
+  @Test
+  void recreateTable() throws Exception {
+    BranchName branch = BranchName.of("recreateTable-main");
+    Key key = Key.of("recreateTable");
+
+    store().create(branch, Optional.empty());
+    // commit just something to have a "real" common ancestor and not "beginning of time", which
+    // means no-common-ancestor
+    Hash ancestor =
+        store()
+            .commit(
+                branch,
+                Optional.empty(),
+                "create table",
+                singletonList(
+                    Put.of(
+                        key,
+                        StringStoreWorker.withStateAndId(
+                            "initial-state", "value", "CONTENT-ID-1"))));
+    assertThat(store().getValue(branch, key)).isEqualTo("initial-state|value@CONTENT-ID-1");
+    assertThat(store().getValue(ancestor, key)).isEqualTo("initial-state|value@CONTENT-ID-1");
+
+    Hash delete =
+        store().commit(branch, Optional.empty(), "drop table", ImmutableList.of(Delete.of(key)));
+    assertThat(store().getValue(branch, key)).isNull();
+    assertThat(store().getValue(delete, key)).isNull();
+
+    Hash recreate =
+        store()
+            .commit(
+                branch,
+                Optional.empty(),
+                "drop table",
+                ImmutableList.of(
+                    Put.of(
+                        key,
+                        StringStoreWorker.withStateAndId(
+                            "recreate-state", "value", "CONTENT-ID-DIFFERENT"))));
+    assertThat(store().getValue(branch, key))
+        .isEqualTo("recreate-state|value@CONTENT-ID-DIFFERENT");
+    assertThat(store().getValue(recreate, key))
+        .isEqualTo("recreate-state|value@CONTENT-ID-DIFFERENT");
+  }
+
+  private static class ReferenceNotFoundFunction {
+
+    final String name;
+    String msg;
+    ThrowingFunction setup;
+    ThrowingFunction function;
+
+    ReferenceNotFoundFunction(String name) {
+      this.name = name;
+    }
+
+    ReferenceNotFoundFunction setup(ThrowingFunction setup) {
+      this.setup = setup;
+      return this;
+    }
+
+    ReferenceNotFoundFunction function(ThrowingFunction function) {
+      this.function = function;
+      return this;
+    }
+
+    ReferenceNotFoundFunction msg(String msg) {
+      this.msg = msg;
+      return this;
+    }
+
+    @Override
+    public String toString() {
+      return name;
+    }
+
+    @FunctionalInterface
+    interface ThrowingFunction {
+      void run(VersionStore<String, String, StringStoreWorker.TestEnum> store)
+          throws VersionStoreException;
+    }
+  }
+
+  static List<ReferenceNotFoundFunction> referenceNotFoundFunctions() {
+    return Arrays.asList(
+        // getCommits()
+        new ReferenceNotFoundFunction("getCommits/branch")
+            .msg("Named reference 'this-one-should-not-exist' not found")
+            .function(s -> s.getCommits(BranchName.of("this-one-should-not-exist"), false)),
+        new ReferenceNotFoundFunction("getCommits/tag")
+            .msg("Named reference 'this-one-should-not-exist' not found")
+            .function(s -> s.getCommits(TagName.of("this-one-should-not-exist"), false)),
+        new ReferenceNotFoundFunction("getCommits/hash")
+            .msg("Commit '12341234123412341234123412341234123412341234' not found")
+            .function(
+                s -> s.getCommits(Hash.of("12341234123412341234123412341234123412341234"), false)),
+        // getValue()
+        new ReferenceNotFoundFunction("getValue/branch")
+            .msg("Named reference 'this-one-should-not-exist' not found")
+            .function(s -> s.getValue(BranchName.of("this-one-should-not-exist"), Key.of("foo"))),
+        new ReferenceNotFoundFunction("getValue/tag")
+            .msg("Named reference 'this-one-should-not-exist' not found")
+            .function(s -> s.getValue(TagName.of("this-one-should-not-exist"), Key.of("foo"))),
+        new ReferenceNotFoundFunction("getValue/hash")
+            .msg("Commit '12341234123412341234123412341234123412341234' not found")
+            .function(
+                s ->
+                    s.getValue(
+                        Hash.of("12341234123412341234123412341234123412341234"), Key.of("foo"))),
+        // getValues()
+        new ReferenceNotFoundFunction("getValues/branch")
+            .msg("Named reference 'this-one-should-not-exist' not found")
+            .function(
+                s ->
+                    s.getValues(
+                        BranchName.of("this-one-should-not-exist"), singletonList(Key.of("foo")))),
+        new ReferenceNotFoundFunction("getValues/tag")
+            .msg("Named reference 'this-one-should-not-exist' not found")
+            .function(
+                s ->
+                    s.getValues(
+                        TagName.of("this-one-should-not-exist"), singletonList(Key.of("foo")))),
+        new ReferenceNotFoundFunction("getValues/hash")
+            .msg("Commit '12341234123412341234123412341234123412341234' not found")
+            .function(
+                s ->
+                    s.getValues(
+                        Hash.of("12341234123412341234123412341234123412341234"),
+                        singletonList(Key.of("foo")))),
+        // getKeys()
+        new ReferenceNotFoundFunction("getKeys/branch")
+            .msg("Named reference 'this-one-should-not-exist' not found")
+            .function(s -> s.getKeys(BranchName.of("this-one-should-not-exist"))),
+        new ReferenceNotFoundFunction("getKeys/tag")
+            .msg("Named reference 'this-one-should-not-exist' not found")
+            .function(s -> s.getKeys(TagName.of("this-one-should-not-exist"))),
+        new ReferenceNotFoundFunction("getKeys/hash")
+            .msg("Commit '12341234123412341234123412341234123412341234' not found")
+            .function(s -> s.getKeys(Hash.of("12341234123412341234123412341234123412341234"))),
+        // assign()
+        new ReferenceNotFoundFunction("assign/branch/ok")
+            .msg("Named reference 'this-one-should-not-exist' not found")
+            .function(
+                s ->
+                    s.assign(
+                        BranchName.of("this-one-should-not-exist"),
+                        Optional.empty(),
+                        s.noAncestorHash())),
+        new ReferenceNotFoundFunction("assign/hash")
+            .msg("Commit '12341234123412341234123412341234123412341234' not found")
+            .function(
+                s ->
+                    s.assign(
+                        BranchName.of("main"),
+                        Optional.empty(),
+                        Hash.of("12341234123412341234123412341234123412341234"))),
+        // delete()
+        new ReferenceNotFoundFunction("delete/branch")
+            .msg("Named reference 'this-one-should-not-exist' not found")
+            .function(s -> s.delete(BranchName.of("this-one-should-not-exist"), Optional.empty())),
+        new ReferenceNotFoundFunction("delete/tag")
+            .msg("Named reference 'this-one-should-not-exist' not found")
+            .function(s -> s.delete(BranchName.of("this-one-should-not-exist"), Optional.empty())),
+        // create()
+        new ReferenceNotFoundFunction("create/hash")
+            .msg("Commit '12341234123412341234123412341234123412341234' not found")
+            .function(
+                s ->
+                    s.create(
+                        BranchName.of("foo"),
+                        Optional.of(Hash.of("12341234123412341234123412341234123412341234")))),
+        // commit()
+        new ReferenceNotFoundFunction("commit/branch")
+            .msg("Named reference 'this-one-should-not-exist' not found")
+            .function(
+                s ->
+                    s.commit(
+                        BranchName.of("this-one-should-not-exist"),
+                        Optional.empty(),
+                        "meta",
+                        singletonList(Delete.of(Key.of("meep"))))),
+        new ReferenceNotFoundFunction("commit/hash")
+            .msg(
+                "Could not find commit '12341234123412341234123412341234123412341234' in reference 'main'.")
+            .function(
+                s ->
+                    s.commit(
+                        BranchName.of("main"),
+                        Optional.of(Hash.of("12341234123412341234123412341234123412341234")),
+                        "meta",
+                        singletonList(Delete.of(Key.of("meep"))))),
+        // transplant()
+        new ReferenceNotFoundFunction("transplant/branch/ok")
+            .msg("Named reference 'this-one-should-not-exist' not found")
+            .function(
+                s ->
+                    s.transplant(
+                        BranchName.of("this-one-should-not-exist"),
+                        Optional.empty(),
+                        singletonList(s.hashOnReference(BranchName.of("main"), Optional.empty())),
+                        Function.identity())),
+        new ReferenceNotFoundFunction("transplant/hash/empty")
+            .msg(
+                "Could not find commit '12341234123412341234123412341234123412341234' in reference 'main'.")
+            .function(
+                s ->
+                    s.transplant(
+                        BranchName.of("main"),
+                        Optional.of(Hash.of("12341234123412341234123412341234123412341234")),
+                        singletonList(Hash.of("12341234123412341234123412341234123412341234")),
+                        Function.identity())),
+        new ReferenceNotFoundFunction("transplant/empty/hash")
+            .msg("Commit '12341234123412341234123412341234123412341234' not found")
+            .function(
+                s ->
+                    s.transplant(
+                        BranchName.of("main"),
+                        Optional.empty(),
+                        singletonList(Hash.of("12341234123412341234123412341234123412341234")),
+                        Function.identity())),
+        // merge()
+        new ReferenceNotFoundFunction("merge/hash/empty")
+            .msg("Commit '12341234123412341234123412341234123412341234' not found")
+            .function(
+                s ->
+                    s.merge(
+                        Hash.of("12341234123412341234123412341234123412341234"),
+                        BranchName.of("main"),
+                        Optional.empty(),
+                        Function.identity())),
+        new ReferenceNotFoundFunction("merge/empty/hash")
+            .msg(
+                "Could not find commit '12341234123412341234123412341234123412341234' in reference 'main'.")
+            .function(
+                s ->
+                    s.merge(
+                        s.noAncestorHash(),
+                        BranchName.of("main"),
+                        Optional.of(Hash.of("12341234123412341234123412341234123412341234")),
+                        Function.identity())));
+  }
+
+  @ParameterizedTest
+  @MethodSource("referenceNotFoundFunctions")
+  void referenceNotFound(ReferenceNotFoundFunction f) throws Exception {
+    if (f.setup != null) {
+      f.setup.run(store());
+    }
+    assertThatThrownBy(() -> f.function.run(store()))
+        .isInstanceOf(ReferenceNotFoundException.class)
+        .hasMessage(f.msg);
+  }
+
+  /** Assigning a branch/tag to a fresh main without any commits didn't work in 0.9.2 */
+  @Test
+  public void assignReferenceToFreshMain()
+      throws ReferenceNotFoundException, ReferenceAlreadyExistsException,
+          ReferenceConflictException {
+    BranchName main = BranchName.of("main");
+    WithHash<Ref> mainRef = store().toRef(main.getName());
+    assertThat(store().getCommits(main, false)).isEmpty();
+    try (Stream<ReferenceInfo<String>> refs = store().getNamedRefs(GetNamedRefsParams.DEFAULT)) {
+      assertThat(refs).extracting(r -> r.getNamedRef().getName()).containsExactly(main.getName());
+    }
+
+    BranchName testBranch = BranchName.of("testBranch");
+    Hash testBranchHash = store().create(testBranch, Optional.empty());
+    store().assign(testBranch, Optional.of(testBranchHash), mainRef.getHash());
+    assertThat(store().toRef(testBranch.getName()).getHash()).isEqualTo(mainRef.getHash());
+
+    TagName testTag = TagName.of("testTag");
+    Hash testTagHash = store().create(testTag, Optional.empty());
+    store().assign(testTag, Optional.of(testTagHash), mainRef.getHash());
+    assertThat(store().toRef(testTag.getName()).getHash()).isEqualTo(mainRef.getHash());
   }
 }
